delete idea_part2/rename_date
